+++ src/context/ItemProvider.tsx
@@ -1,0 +1,43 @@
+import { createContext, useMemo, useState } from "react";
+import { generateItems } from "../utils";
+
+interface Item {
+  id: number;
+  name: string;
+  category: string;
+  price: number;
+}
+
+interface ContextValue {
+  items: Item[];
+  addItems: () => void;
+}
+
+export const ItemContext = createContext<ContextValue | null>(null);
+
+type Props = {
+  children: React.ReactNode;
+};
+
+export const ItemProvider = ({ children }: Props) => {
+  const [items, setItems] = useState(generateItems(1000));
+
+  const addItems = () => {
+    setItems((prevItems) => [
+      ...prevItems,
+      ...generateItems(1000, prevItems.length),
+    ]);
+  };
+
+  const contextValue: ContextValue = useMemo(
+    () => ({
+      items,
+      addItems,
+    }),
+    [items, addItems],
+  );
+
+  return (
+    <ItemContext.Provider value={contextValue}>{children}</ItemContext.Provider>
+  );
+};
